+++ examples/resources/sdwan_class_map_policy_object/resource.tf
@@ -1,8 +1,4 @@
 resource "sdwan_class_map_policy_object" "example" {
-  name = "Example"
-  entries = [
-    {
-      queue = 2
-    }
-  ]
+  name  = "Example"
+  queue = 2
 }