+++ internal/provider/resource_sdwan_class_map_policy_object_test.go
@@ -31,22 +31,21 @@ func TestAccSdwanClassMapPolicyObject(t *testing.T) {
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccSdwanClassMapPolicyObjectConfig_all(),
+				Config: testAccSdwanClassMapPolicyObjectConfig,
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr("sdwan_class_map_policy_object.test", "entries.0.queue", "2"),
+					resource.TestCheckResourceAttr("sdwan_class_map_policy_object.test", "name", "Example"),
+					resource.TestCheckResourceAttr("sdwan_class_map_policy_object.test", "queue", "2"),
 				),
 			},
 		},
 	})
 }
 
-func testAccSdwanClassMapPolicyObjectConfig_all() string {
-	return `
-	resource "sdwan_class_map_policy_object" "test" {
-		name = "TF_TEST_ALL"
-		entries = [{
-			queue = 2
-		}]
-	}
-	`
+const testAccSdwanClassMapPolicyObjectConfig = `
+
+
+resource "sdwan_class_map_policy_object" "test" {
+	name = "Example"
+	queue = 2
 }
+`