+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add `set_parameters` to `sdwan_acl_policy_definition` resource and data source
 - BREAKING CHANGE: Remove `dscp` and `next_hop` attributes of `sdwan_acl_policy_definition` resource and data source
 - BREAKING CHANGE: Rename `layer2cos` attribute of `sdwan_rewrite_rule_policy_definition` resource and data source to `layer2_cos`
+- BREAKING CHANGE: Remove `entries` attribute of `sdwan_class_map_policy_object` resource and data source
 
 ## 0.2.1